Create a project for graph optimizer.
Move optimizer related code to the folder optimizer.
