+++ include/onnxruntime/core/optimizer/graph_transformer.h
@@ -5,7 +5,7 @@
 
 #include "core/common/common.h"
 #include "core/graph/graph_viewer.h"
-#include "core/graph/rewrite_rule.h"
+#include "core/optimizer/rewrite_rule.h"
 
 namespace onnxruntime {
