+++ onnxruntime/core/framework/insert_cast_transformer.h
@@ -3,7 +3,7 @@
 
 #pragma once
 #include "core/graph/graph_viewer.h"
-#include "core/graph/graph_transformer.h"
+#include "core/optimizer/graph_transformer.h"
 #include "core/common/common.h"
 #include "core/framework/op_kernel.h"
