+++ onnxruntime/core/framework/session_state_initializer.cc
@@ -10,8 +10,8 @@
 #include "core/common/logging/logging.h"
 
 #include "core/graph/graph_viewer.h"
-#include "core/graph/graph_transformer.h"
-#include "core/graph/graph_transformer_mgr.h"
+#include "core/optimizer/graph_transformer.h"
+#include "core/optimizer/graph_transformer_mgr.h"
 
 #include "core/framework/graph_partitioner.h"
 #include "core/framework/insert_cast_transformer.h"
