+++ onnxruntime/core/optimizer/conv_activation_fusion.cc
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/graph/initializer.h"
-#include "core/graph/conv_activation_fusion.h"
-#include "core/graph/graph_utils.h"
 #include <deque>
+#include "core/graph/graph_utils.h"
+#include "core/optimizer/initializer.h"
+#include "core/optimizer/conv_activation_fusion.h"
 
 using namespace onnx;
 using namespace ::onnxruntime::common;
