+++ onnxruntime/core/optimizer/conv_activation_fusion.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "core/graph/graph_transformer.h"
+#include "core/optimizer/graph_transformer.h"
 
 namespace onnxruntime {
