+++ onnxruntime/core/optimizer/conv_add_fusion.h
@@ -2,13 +2,14 @@
 // Licensed under the MIT License.
 
 #pragma once
-#include "core/graph/graph_transformer.h"
+
+#include "core/optimizer/graph_transformer.h"
 
 namespace onnxruntime {
 
-class ConvMulFusion : public onnxruntime::GraphTransformer {
- public:
-  ConvMulFusion() noexcept : onnxruntime::GraphTransformer("ConvMulFusion", "Fusing Mul into Conv") {}
+class ConvAddFusion : public onnxruntime::GraphTransformer {
+public:
+  ConvAddFusion() noexcept : onnxruntime::GraphTransformer("ConvAddFusion", "Fusing Add into Conv") {}
   Status Apply(onnxruntime::Graph& graph, bool& modified) const override;
 };
