+++ onnxruntime/core/optimizer/conv_mul_fusion.h
@@ -2,14 +2,13 @@
 // Licensed under the MIT License.
 
 #pragma once
-
-#include "core/graph/graph_transformer.h"
+#include "core/optimizer/graph_transformer.h"
 
 namespace onnxruntime {
 
-class UnsqueezeElimination : public onnxruntime::GraphTransformer {
+class ConvMulFusion : public onnxruntime::GraphTransformer {
  public:
-  UnsqueezeElimination() noexcept : onnxruntime::GraphTransformer("EliminateUnsqueeze", "Eliminate unsequeeze node") {}
+  ConvMulFusion() noexcept : onnxruntime::GraphTransformer("ConvMulFusion", "Fusing Mul into Conv") {}
   Status Apply(onnxruntime::Graph& graph, bool& modified) const override;
 };
