+++ onnxruntime/core/optimizer/graph_transformer_mgr.cc
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/graph/graph_transformer_mgr.h"
+#include "core/optimizer/graph_transformer_mgr.h"
 using namespace onnxruntime;
 using namespace ::onnxruntime::common;
