+++ onnxruntime/core/session/inference_session.cc
@@ -16,8 +16,8 @@
 #include "core/common/logging/logging.h"
 #include "core/common/task_thread_pool.h"
 #include "core/graph/graph_viewer.h"
-#include "core/graph/graph_transformer.h"
-#include "core/graph/graph_transformer_mgr.h"
+#include "core/optimizer/graph_transformer.h"
+#include "core/optimizer/graph_transformer_mgr.h"
 #include "core/graph/graph_utils.h"
 #include "core/graph/model.h"
 #include "core/framework/allocatormgr.h"
