+++ onnxruntime/test/ir/graph_transform_test.cc
@@ -4,17 +4,16 @@
 #include "core/session/inference_session.h"
 #include "core/graph/graph_viewer.h"
 #include "core/graph/model.h"
-#include "core/graph/graph_transformer.h"
-#include "core/graph/identity_elimination.h"
-#include "core/graph/unsqueeze_elimination.h"
-#include "core/graph/conv_bn_fusion.h"
-#include "core/graph/conv_mul_fusion.h"
-#include "core/graph/conv_add_fusion.h"
-#include "core/graph/conv_activation_fusion.h"
-#include "core/graph/matmul_add_fusion.h"
-#include "core/graph/gemm_activation_fusion.h"
+#include "core/optimizer/graph_transformer.h"
+#include "core/optimizer/identity_elimination.h"
+#include "core/optimizer/unsqueeze_elimination.h"
+#include "core/optimizer/conv_bn_fusion.h"
+#include "core/optimizer/conv_mul_fusion.h"
+#include "core/optimizer/conv_add_fusion.h"
+#include "core/optimizer/conv_activation_fusion.h"
+#include "core/optimizer/matmul_add_fusion.h"
+#include "core/optimizer/gemm_activation_fusion.h"
 #include "core/platform/env.h"
-
 #include "test/capturing_sink.h"
 #include "test/test_environment.h"
 #include "gtest/gtest.h"
